Run personal JSON sync during app initialization
