+++ app/src/main/java/at/plankt0n/streamplay/StreamPlayApplication.kt
@@ -1,11 +1,28 @@
 package at.plankt0n.streamplay
 
 import android.app.Application
+import android.util.Log
+import androidx.preference.PreferenceManager
 import at.plankt0n.streamplay.helper.CrashHandler
+import at.plankt0n.streamplay.helper.StationImportHelper
+import kotlinx.coroutines.runBlocking
 
 class StreamPlayApplication : Application() {
     override fun onCreate() {
         super.onCreate()
+
+        val prefs = PreferenceManager.getDefaultSharedPreferences(this)
+        if (prefs.getBoolean(Keys.PREF_PERSONAL_SYNC_STARTUP, false)) {
+            val url = prefs.getString(Keys.PREF_PERSONAL_SYNC_URL, Keys.DEFAULT_PERSONAL_SYNC_URL) ?: ""
+            Log.i("StreamPlay", "Syncing personal JSON at startup")
+            runBlocking {
+                try {
+                    StationImportHelper.importStationsFromUrl(this@StreamPlayApplication, url, true)
+                } catch (e: Exception) {
+                    Log.e("StreamPlay", "Startup sync failed: ${e.message}")
+                }
+            }
+        }
 
         Thread.setDefaultUncaughtExceptionHandler(CrashHandler(this))
     }
